requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
ebooklib>=0.18
beautifulsoup4>=4.12.0
requests>=2.31.0
google-generativeai>=0.3.0
urllib3>=2.0.0
tqdm>=4.66.0
pyinstaller>=6.0.0
DrissionPage>=4.0.0
